+++ examples/src/index.js
@@ -5,6 +5,8 @@ import Counter, {MultiCounter} from './Counter'
 import BMICalculator from './BMICalculator'
 import CombineEventsWithTest from './CombineEventsWithTest'
 import PremountEventOrderTest from './PremountEventOrderTest'
+import UnmountErrorTest from './UnmountErrorTest'
+import TeardownTest from './TeardownTest'
 
 storiesOf('Counter', module)
   .add('default', () =>
@@ -16,3 +18,5 @@ storiesOf('BMI Calculator', module).add('default', () => <BMICalculator />)
 storiesOf('Testing', module)
   .add('CombineEventsWithTest', () => <CombineEventsWithTest />)
   .add('PremountEventOrderTest', () => <PremountEventOrderTest />)
+  .add('UnmountErrorTest', () => <UnmountErrorTest />)
+  .add('TeardownTest', () => <TeardownTest />)